Tests for SnippetEmbedHandler
