+++ tests/testapp/templates/testapp/advert_snippet_embed.html
@@ -1,0 +1,1 @@
+<div>{{ object.text }}</div>
