+++ tests/testapp/templates/testapp/advert_snippet_link.html
